# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AgentoGo API",
        "description": "API REST multi-tenant con autenticación, auditoría y gestión de negocios",
        "contact": {
            "name": "AgentoGo Support",
            "email": "support@agentogo.com"
        },
        "version": "0.3.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor de desarrollo"
        }
    ],
    "paths": {
...
        "/businesses/{business_id}/clients": {
            "get": {
                "tags": [
                    "Client"
                ],
                "summary": "Listar clientes del negocio",
                "operationId": "567ea98be06efeac8427f98f2be17acb",
                "responses": {
                    "200": {
                        "description": "Lista de clientes con paginación"
                    },
                    "403": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error403"
                                },
                                "example": {
                                    "code": 403,
                                    "message": "No tiene permisos para realizar esta acción."
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Client"
                ],
                "summary": "Crear un cliente",
                "operationId": "d7ee991434d8ac73abd9844d6016d34c",
                "responses": {
                    "201": {
                        "description": "Cliente creado"
                    },
                    "403": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error403"
                                },
                                "example": {
                                    "code": 403,
                                    "message": "No tiene permisos para realizar esta acción."
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validación fallida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                },
                                "example": {
                                    "message": "The given data was invalid.",
                                    "errors": {
                                        "name": [
                                            "El nombre es obligatorio."
                                        ],
                                        "email": [
                                            "El email debe ser una dirección válida."
                                        ]
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/businesses/{business_id}/clients/{client_id}": {
            "get": {
                "tags": [
                    "Client"
                ],
                "summary": "Obtener un cliente",
                "operationId": "cee977ae365a631dc4e17a83e617e634",
                "responses": {
                    "200": {
                        "description": "Detalle del cliente"
                    },
                    "403": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error403"
                                },
                                "example": {
                                    "code": 403,
                                    "message": "No tiene permisos para realizar esta acción."
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "No encontrado"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Client"
                ],
                "summary": "Actualizar un cliente",
                "operationId": "bf4f775e93f353219533d0c25db7f40d",
                "responses": {
                    "200": {
                        "description": "Cliente actualizado"
                    },
                    "403": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error403"
                                },
                                "example": {
                                    "code": 403,
                                    "message": "No tiene permisos para realizar esta acción."
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validación fallida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                },
                                "example": {
                                    "message": "The given data was invalid.",
                                    "errors": {
                                        "name": [
                                            "El nombre es obligatorio."
                                        ],
                                        "email": [
                                            "El email debe ser una dirección válida."
                                        ]
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Client"
                ],
                "summary": "Eliminar un cliente",
                "operationId": "c544822adcae773b1ef73d2cbf497258",
                "responses": {
                    "204": {
                        "description": "Cliente eliminado"
                    },
                    "403": {
                        "description": "No autorizado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error403"
                                },
                                "example": {
                                    "code": 403,
                                    "message": "No tiene permisos para realizar esta acción."
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Error403": {
                "properties": {
                    "code": {
                        "type": "integer",
                        "example": 403
                    },
                    "message": {
                        "type": "string",
                        "example": "No autorizado"
                    }
                },
                "type": "object",
                "example": {
                    "code": 403,
                    "message": "No tiene permisos para realizar esta acción."
                }
            },
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "description": "Mapa de errores por campo",
                        "type": "object"
                    }
                },
                "type": "object",
                "example": {
                    "message": "The given data was invalid.",
                    "errors": {
                        "name": [
                            "El campo nombre es requerido."
                        ],
                        "email": [
                            "El email debe ser una dirección válida."
                        ]
                    }
                }
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token de autenticación Sanctum",
                "bearerFormat": "token",
                "scheme": "bearer"
            }
        }
    },
...
}